fix: merge-create-filter
